added permissions and absolute paths to delete teacher page

+++ admin/lehrer/delete.php
@@ -12,8 +12,8 @@
 <html lang="de-DE" prefix="og: https://ogp.me/ns#" xmlns:og="http://opengraphprotocol.org/schema/">
     <head>
         <?php
-
-            include_once "./../../sites/head.html"
+            $root = realpath($_SERVER["DOCUMENT_ROOT"]);
+            include_once "$root/admin/sites/head.html";
 
         ?>
         <title>Lehrer löschen - Admin Panel - Friedrich-Gymnasium Luckenwalde</title>
@@ -21,7 +21,16 @@
     <body>
         <?php
 
-            include_once "./../../sites/header.html"
+            include_once "$root/admin/sites/header.html";
+
+            include_once "$root/admin/sites/permissions.php";
+
+            include_once "$root/admin/no-permission.html";
+
+            if($lehrer_all == 0){
+                echo("<script>$('.no_perm').show();</script>");
+                $disabled = true;
+            };
 
         ?>
 
@@ -38,7 +47,7 @@
             if ($conn->connect_error) {
                 die("Connection failed: " . $conn->connect_error);
             }
-            $insert = mysqli_query($conn, "DELETE FROM lehrer WHERE id='{$id}'");
+            if ($disabled==false){$insert = mysqli_query($conn, "DELETE FROM lehrer WHERE id='{$id}'");}
             if ($insert) {
                 echo '<script type="text/javascript">window.location = "/admin/lehrer/"</script>';
             }
